Separated Token Handling into it's own package to clean up main code.
