+++ bin/TokenHandler/TokenType.h
@@ -1,0 +1,6 @@
+#ifndef LISPCOMPILER_TOKENTYPE_H
+#define LISPCOMPILER_TOKENTYPE_H
+
+enum TokenType {NumberToken, ParenToken, NameToken};
+
+#endif //LISPCOMPILER_TOKENTYPE_H
