+++ bin/TokenHandler/Tokenizer.h
@@ -1,0 +1,10 @@
+#ifndef LISPCOMPILER_TOKENIZER_H
+#define LISPCOMPILER_TOKENIZER_H
+
+
+#include <vector>
+#include "Token.h"
+
+std::vector<Token> tokenizer(std::string _inputCode);
+
+#endif //LISPCOMPILER_TOKENIZER_H
